Exchange of doubles to floats for optimized performance of THE task

+++ PWGCF/FEMTOSCOPY/FemtoDream/AddTaskFemtoDream.C
@@ -112,11 +112,11 @@ AliAnalysisTask* AddTaskFemtoDream(bool isMC=false, TString CentEst="kInt7",
 	AntiCascadeCuts->SetPDGCodeBach(-211);
 
 	std::vector<int> PDGParticles ={2212,2212,3122,3122,3312,3312};
-	std::vector<double> ZVtxBins = {-10,-8,-6,-4,-2,0,2,4,6,8,10};
+	std::vector<float> ZVtxBins = {-10,-8,-6,-4,-2,0,2,4,6,8,10};
 	std::vector<int> MultBins = {0,4,8,12,16,20,24,28,32,36,40,44,48,52,56,60,64,80};
 	std::vector<int> NBins= {750,750,150,150,150,150,750,150,150,150,150,150,150,150,150,150,150,150,150,150,150};
-	std::vector<double> kMin= {0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.};
-	std::vector<double> kMax= {3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.};
+	std::vector<float> kMin= {0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.,0.};
+	std::vector<float> kMax= {3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.,3.};
 	AliFemtoDreamCollConfig *config=new AliFemtoDreamCollConfig("Femto","Femto");
 	config->SetMultBinning(true);
 	config->SetZBins(ZVtxBins);
