+++ PWGCF/FEMTOSCOPY/FemtoDream/AliFemtoDreamCollConfig.h
@@ -18,16 +18,16 @@ class AliFemtoDreamCollConfig : public TNamed {
   AliFemtoDreamCollConfig(const char *name, const char *title);
   virtual ~AliFemtoDreamCollConfig();
   void SetMultBinning(bool doIt){fMultBinning=doIt;};
-  void SetZBins(std::vector<double> ZBins);
+  void SetZBins(std::vector<float> ZBins);
   void SetMultBins(std::vector<int> MultBins);
   void SetPDGCodes(std::vector<int> PDGCodes);
   void SetNBinsHist(std::vector<int> NBins);
-  void SetMinKRel(std::vector<double> minKRel);
-  void SetMaxKRel(std::vector<double> maxKRel);
+  void SetMinKRel(std::vector<float> minKRel);
+  void SetMaxKRel(std::vector<float> maxKRel);
   void SetMixingDepth(int MixingDepth){fMixingDepth=MixingDepth;};
 
   bool GetDoMultBinning() {return fMultBinning;};
-  std::vector<double> GetZVtxBins();
+  std::vector<float> GetZVtxBins();
   int GetNZVtxBins(){return (fZVtxBins->GetEntries()-1);};
   std::vector<int> GetMultBins();
   int GetNMultBins(){return fMultBins->GetEntries();};
@@ -35,8 +35,8 @@ class AliFemtoDreamCollConfig : public TNamed {
   int GetNParticles() {return fPDGParticleSpecies->GetEntries();};
   int GetNParticleCombinations();
   std::vector<int> GetNBinsHist();
-  std::vector<double> GetMinKRel();
-  std::vector<double> GetMaxKRel();
+  std::vector<float> GetMinKRel();
+  std::vector<float> GetMaxKRel();
   int GetMixingDepth(){return fMixingDepth;};
  private:
   bool fMultBinning;            //
